pyproject.toml
[tool.poetry]
name = "berserk-backend"
version = "0.1.0"
description = "Backend for berserk, the advanced tracking system."
authors = ["shashwat2003"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.7"
pyyaml = "^6.0.1"
psycopg2-binary = "^2.9.9"
djangorestframework = "^3.15.2"
pre-commit = "^3.8.0"
uuid6 = "^2024.7.10"
drf-spectacular = "^0.27.2"
pydantic = "^2.9.2"
django-cors-headers = "^4.6.0"

[tool.poetry.group.dev.dependencies]
mypy = "^1.11.2"
django-sslserver-v2 = "^1.0"
django-stubs = "^5.0.4"
types-psycopg2 = "^2.9.21.20240819"
djangorestframework-stubs = {extras = ["compatible-mypy"], version = "^3.15.1"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
